Type fixes

+++ freecad/gridfinity_workbench/junction_screws.py
@@ -28,29 +28,29 @@ def holes_shape(
     top_z: fc.Units.Quantity,
     geometry: GridfinityLayoutGeometry,
 ) -> Part.Shape | None:
-    through_depth = top_z + 0.1 * fc.Units.Quantity("1 mm")
+    through_depth = float(top_z + 0.1 * fc.Units.Quantity("1 mm"))
     cutters: list[Part.Shape] = []
     for x, y in iter_supported_junctions(geometry):
         through = Part.makeCylinder(
-            junction_screws.screw_diameter / 2,
+            float(junction_screws.screw_diameter / 2),
             through_depth,
-            fc.Vector(x, y, top_z),
+            fc.Vector(x, y, float(top_z)),
             fc.Vector(0, 0, -1),
         )
         counterbore = Part.makeCylinder(
-            junction_screws.counterbore_diameter / 2,
-            junction_screws.counterbore_depth,
-            fc.Vector(x, y, top_z),
+            float(junction_screws.counterbore_diameter / 2),
+            float(junction_screws.counterbore_depth),
+            fc.Vector(x, y, float(top_z)),
             fc.Vector(0, 0, -1),
         )
-        transition_height = (
-            junction_screws.counterbore_diameter - junction_screws.screw_diameter
-        ) / 2
+        transition_height = float(
+            (junction_screws.counterbore_diameter - junction_screws.screw_diameter) / 2
+        )
         transition = Part.makeCone(
-            junction_screws.counterbore_diameter / 2,
-            junction_screws.screw_diameter / 2,
+            float(junction_screws.counterbore_diameter / 2),
+            float(junction_screws.screw_diameter / 2),
             transition_height,
-            fc.Vector(x, y, top_z - junction_screws.counterbore_depth),
+            fc.Vector(x, y, float(top_z - junction_screws.counterbore_depth)),
             fc.Vector(0, 0, -1),
         )
         cutters.extend([through, counterbore, transition])
@@ -65,29 +65,25 @@ def stubs_shape(
     bottom_z: fc.Units.Quantity,
     geometry: GridfinityLayoutGeometry,
 ) -> Part.Shape | None:
-    stub_diameter = junction_screws.screw_diameter - screw_stubs.clearance
-    counterbore_diameter = junction_screws.counterbore_diameter - screw_stubs.clearance
+    stub_diameter = float(junction_screws.screw_diameter - screw_stubs.clearance)
+    counterbore_diameter = float(junction_screws.counterbore_diameter - screw_stubs.clearance)
     transition_height = (counterbore_diameter - stub_diameter) / 2
-    if (
-        float(stub_diameter) <= 0
-        or float(counterbore_diameter) <= 0
-        or float(transition_height) <= 0
-    ):
+    if stub_diameter <= 0 or counterbore_diameter <= 0 or transition_height <= 0:
         return None
 
     stubs: list[Part.Shape] = []
     for x, y in iter_supported_junctions(geometry):
         counterbore = Part.makeCylinder(
             counterbore_diameter / 2,
-            junction_screws.counterbore_depth,
-            fc.Vector(x, y, bottom_z),
+            float(junction_screws.counterbore_depth),
+            fc.Vector(x, y, float(bottom_z)),
             fc.Vector(0, 0, -1),
         )
         transition = Part.makeCone(
             counterbore_diameter / 2,
             stub_diameter / 2,
             transition_height,
-            fc.Vector(x, y, bottom_z - junction_screws.counterbore_depth),
+            fc.Vector(x, y, float(bottom_z - junction_screws.counterbore_depth)),
             fc.Vector(0, 0, -1),
         )
         stubs.extend([counterbore, transition])
